working on edit for todo list

+++ src/lib/actions.tsx
@@ -1,6 +1,7 @@
 "use server";
 import prisma from "../../prisma/client";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
 
 export const addToDo = async (formData: FormData) => {
@@ -17,7 +18,19 @@ export const addToDo = async (formData: FormData) => {
   }
 };
 
-export const deleteToDo = async ({ id }: { id: string }) => {
-  await prisma.toDo.delete({ where: { id } });
-  revalidatePath(`/`);
+export const deleteToDo = async ({
+  id,
+  userId,
+}: {
+  id: string;
+  userId: string;
+}) => {
+  const { getUser } = getKindeServerSession();
+  const user = await getUser();
+  if (user.id === userId) {
+    await prisma.toDo.delete({ where: { id } });
+    redirect(`/`);
+  } else {
+    alert("No naughty boys")
+  }
 };
